move errStarterFixture into internal/testutil
Moves the errStarterFixture into its own package in
`internal/testutil/fixture/errstarter`
